requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
Cerberus==1.3.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
environs==9.3.2
kombu==5.1.0
marshmallow==3.13.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
serpy==0.3.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5